requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
bitarray==3.4.3
certifi==2025.6.15
charset-normalizer==3.4.2
ckzg==2.1.1
cytoolz==1.0.1
eth-abi==5.2.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
frozenlist==1.5.0
hexbytes==1.3.1
idna==3.10
multidict==6.1.0
packaging==25.0
parsimonious==0.10.0
propcache==0.2.0
py-solc-x==2.0.4
pycryptodome==3.23.0
pydantic==2.10.6
pydantic-core==2.27.2
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.4
rlp==4.1.0
toolz==1.0.0
types-requests==2.32.0.20241016
typing-extensions==4.13.2
urllib3==2.2.3
web3==7.12.0
websockets==13.1
yarl==1.15.2
